Fix parseInt -> parseFloat

+++ src/zepp_player/device_state/NumberDeviceState.ts
@@ -12,7 +12,7 @@ export class NumberDeviceState extends DeviceStateEntry<number> {
     }
 
     setValue(value: number) {
-        this.value = parseInt(value as any);
+        this.value = parseFloat(value as any);
     }
 
     getNumber(player: ZeppPlayer): number {
